+++ actionmailer/lib/action_mailer/parameterized.rb
@@ -122,13 +122,6 @@ module ActionMailer
         end
     end
 
-    class DeliveryJob < ActionMailer::DeliveryJob # :nodoc:
-      def perform(mailer, mail_method, delivery_method, params, *args)
-        mailer.constantize.with(params).public_send(mail_method, *args).send(delivery_method)
-      end
-      ruby2_keywords(:perform) if respond_to?(:ruby2_keywords, true)
-    end
-
     class MessageDelivery < ActionMailer::MessageDelivery # :nodoc:
       def initialize(mailer_class, action, params, *args)
         super(mailer_class, action, *args)
@@ -148,25 +141,11 @@ module ActionMailer
           if processed?
             super
           else
-            job = delivery_job_class
-
-            if job <= MailDeliveryJob
-              job.set(options).perform_later(
-                @mailer_class.name, @action.to_s, delivery_method.to_s, params: @params, args: @args)
-            else
-              job.set(options).perform_later(
-                @mailer_class.name, @action.to_s, delivery_method.to_s, @params, *@args)
-            end
+            @mailer_class.delivery_job.set(options).perform_later(
+              @mailer_class.name, @action.to_s, delivery_method.to_s, params: @params, args: @args)
           end
         end
 
-        def delivery_job_class
-          if @mailer_class.delivery_job <= MailDeliveryJob
-            @mailer_class.delivery_job
-          else
-            Parameterized::DeliveryJob
-          end
-        end
     end
   end
 end